Revert "Fix color related issue"
This reverts commit 8fbf915703f23bdac8ea3d269b6012cec12ef0b1.

+++ Assignment2/src/academy/pocu/comp2500/assignment2/Calendar.java
@@ -4,7 +4,7 @@ public class Calendar extends Product {
     private CalendarType type;
 
     public Calendar(CalendarType type, ShippingMethod shippingMethod) {
-        super(type.getSize(), new Color((short) 0xff, (short) 0xff, (short) 0xff), type.getPrice(), shippingMethod);
+        super(type.getSize(), new Color(0xff, 0xff, 0xff), type.getPrice(), shippingMethod);
         String name = String.format("%s Calendar", type.getType());
         super.setName(name);
         this.type = type;
